pyproject.toml
[project]
name = "forjate"
version = "0.1.0"
description = ""
authors = [
    {name = "Sebas Etchevest",email = "setchevest@gmail.com"}
]
package-mode = false
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pyyaml (>=6.0.2,<7.0.0)",
    "jinja2 (>=3.1.6,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
